docs: cut over trees-dev to canonical demos
