+++ apps/docs/app/trees-dev/cookies.ts
@@ -1,4 +1,3 @@
-export const FILE_TREE_COOKIE_VERSION = '1';
+export const FILE_TREE_COOKIE_VERSION = '2';
 export const FILE_TREE_COOKIE_VERSION_NAME = 'ft-controls-version';
 export const FILE_TREE_COOKIE_FLATTEN = 'ft-flatten-empty';
-export const FILE_TREE_COOKIE_LAZY = 'ft-lazy-loader';
