+++ apps/docs/app/trees-dev/drag-and-drop/page.tsx
@@ -1,19 +1,66 @@
+import type { FileTreeOptions } from '@pierre/trees';
 import { preloadFileTree } from '@pierre/trees/ssr';
 
 import { readSettingsCookies } from '../_components/readSettingsCookies';
-import { sharedDemoFileTreeOptions, sharedDemoStateConfig } from '../demo-data';
-import { DragAndDropDemoClient } from './DragAndDropDemoClient';
+import { PathStoreDragAndDropDemoClient } from '../path-store-drag-and-drop/PathStoreDragAndDropDemoClient';
+import { createPresortedPreparedInput } from '../path-store-powered/createPresortedPreparedInput';
 
-export default async function DragAndDropPage() {
-  const { flattenEmptyDirectories, useLazyDataLoader } =
-    await readSettingsCookies();
-  const fileTreeOptions = {
-    ...sharedDemoFileTreeOptions,
+const DRAG_AND_DROP_DEMO_PATHS = [
+  'assets/images/social/banner.png',
+  'assets/images/social/logo.png',
+  'docs/guides/faq.md',
+  'docs/guides/getting-started.md',
+  'src/components/Button.tsx',
+  'src/lib/theme.ts',
+  'src/lib/utils.ts',
+  'src/index.ts',
+  ...Array.from(
+    { length: 40 },
+    (_, index) => `workspace/demo-${String(index).padStart(2, '0')}.ts`
+  ),
+  'package.json',
+  'README.md',
+] as const;
+
+const DRAG_AND_DROP_PREPARED_INPUT = createPresortedPreparedInput(
+  DRAG_AND_DROP_DEMO_PATHS
+);
+const TREE_HEADER_HTML =
+  '<div data-path-store-demo-header style="align-items:center;display:flex;gap:12px;padding:8px 12px"><strong>Drag and drop demo</strong><span>Pointer + touch moves on the canonical tree</span></div>';
+
+export default async function TreesDevDragAndDropPage() {
+  const { flattenEmptyDirectories } = await readSettingsCookies();
+  const sharedOptions: Omit<FileTreeOptions, 'dragAndDrop' | 'id'> = {
+    composition: {
+      header: {
+        html: TREE_HEADER_HTML,
+      },
+    },
     flattenEmptyDirectories,
-    useLazyDataLoader,
+    fileTreeSearchMode: 'hide-non-matches',
+    initialExpandedPaths: [
+      'assets/images/social/',
+      'docs/guides/',
+      'src/',
+      'src/lib/',
+      'workspace/',
+    ],
+    paths: DRAG_AND_DROP_PREPARED_INPUT.paths,
+    preparedInput: DRAG_AND_DROP_PREPARED_INPUT,
+    search: true,
+    viewportHeight: 460,
   };
 
-  const mainSsr = preloadFileTree(fileTreeOptions, sharedDemoStateConfig);
+  const payload = preloadFileTree({
+    ...sharedOptions,
+    dragAndDrop: true,
+    id: 'trees-drag-and-drop',
+  });
 
-  return <DragAndDropDemoClient preloadedFileTreeHtml={mainSsr.shadowHtml} />;
+  return (
+    <PathStoreDragAndDropDemoClient
+      containerHtml={payload.html}
+      sharedOptions={sharedOptions}
+    />
+  );
 }
