+++ apps/docs/app/trees-dev/path-store-git-status/PathStoreGitStatusDemoClient.tsx
@@ -1,9 +1,9 @@
 'use client';
 
 import {
-  PathStoreFileTree,
-  type PathStoreFileTreeOptions,
-} from '@pierre/trees/path-store';
+  FileTree as PathStoreFileTree,
+  type FileTreeOptions as PathStoreFileTreeOptions,
+} from '@pierre/trees';
 import { useEffect, useMemo, useRef } from 'react';
 
 import { ExampleCard } from '../_components/ExampleCard';
@@ -20,12 +20,12 @@ export function PathStoreGitStatusDemoClient({
 }: PathStoreGitStatusDemoClientProps) {
   const mountRef = useRef<HTMLDivElement | null>(null);
   const treeRef = useRef<PathStoreFileTree | null>(null);
-  const { gitStatus, controls } = useGitStatusControls('path-store');
+  const { gitStatus, controls } = useGitStatusControls('canonical');
   const initialGitStatusRef = useRef(gitStatus);
   const options = useMemo<PathStoreFileTreeOptions>(
     () => ({
       ...sharedOptions,
-      id: 'pst-phase9-git-status',
+      id: 'trees-git-status',
     }),
     [sharedOptions]
   );
@@ -63,22 +63,17 @@ export function PathStoreGitStatusDemoClient({
   return (
     <div className="space-y-6">
       <header className="space-y-2">
-        <p className="text-muted-foreground text-xs font-semibold tracking-wide uppercase">
-          Path-store lane · provisional
-        </p>
-        <h1 className="text-2xl font-bold">Path-store Git Status</h1>
+        <h1 className="text-2xl font-bold">Git Status</h1>
         <p className="text-muted-foreground max-w-3xl text-sm leading-6">
-          Phase 9 restores git-status decoration without reviving the legacy
-          core feature seam: changed files render A/M/D markers, folders with
-          changed descendants render the shared dot indicator, and the controls
-          below call <code>setGitStatus()</code> on the live tree instance so
-          the row attrs and status slot update without recreating the tree.
+          Changed files render A/M/D markers, folders with changed descendants
+          keep their shared dot indicator, and the controls below call
+          <code>setGitStatus()</code> on the live tree without recreating it.
         </p>
       </header>
 
       <ExampleCard
-        title="Git-status tree proof"
-        description="Toggle git-status off or swap between the two demo sets. The hydrated path-store tree keeps the same instance while the shared status slot and semantic row attrs update in place."
+        title="Git-status tree"
+        description="Toggle git status off or swap between the two demo sets. The hydrated tree keeps the same instance while the shared status slot and semantic row attrs update in place."
         controls={controls}
       >
         <div
