+++ apps/docs/app/trees-dev/path-store-powered/createPresortedPreparedInput.ts
@@ -1,11 +1,11 @@
-import type { PathStoreFileTreeOptions } from '@pierre/trees/path-store';
+import type { FileTreeOptions as PathStoreFileTreeOptions } from '@pierre/trees';
 
 type PathStorePreparedInput = NonNullable<
   PathStoreFileTreeOptions['preparedInput']
 >;
 
 // This helper exists for demos whose input is already ordered according to the
-// same path-store sort semantics the live tree will use. It does not sort or
+// same tree sort semantics the live tree will use. It does not sort or
 // validate; callers must pass presorted paths.
 export function createPresortedPreparedInput(
   paths: readonly string[]
